CORCI-918 build: fix test_post_always_node.sh
Skip-bullseye: false
Skip-run_test: true
Skip-coverity-test: true
Skip-func-test: true
Skip-func-hw-test: true
Skip-test-centos-rpms: true
Skip-scan-centos-rpms: true
Skip-build-leap-icc: true
Quick-build: true
Skip-checkpatch: true

+++ ci/unit/test_post_always_node.sh
@@ -5,7 +5,9 @@ set -uex
 cd "$DAOS_BASE"
 mkdir run_test.sh
 mkdir vm_test
-mv nlt-errors.json vm_test/
+if [ -e nlt-errors.json ]; then
+  mv nlt-errors.json vm_test/
+fi
 if ls /tmp/daos*.log > /dev/null; then
   mv /tmp/daos*.log run_test.sh/
 fi
